requirements.txt
pymongo
python-dotenv
customtkinter
tk
